requirements.txt
openai
python-dotenv
numpy
scikit-learn
plotly
pandas
matplotlib
